Exercice 4

+++ init.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS `orders` (
   `name` VARCHAR(255) NOT NULL,
   `isin` VARCHAR(255) NOT NULL,
   `amount` INT NOT NULL,
-  `price` INT DEFAULT 0,
+  `price` DECIMAL(10, 2) DEFAULT 0,
   `state` INT DEFAULT 0,
   PRIMARY KEY (`id`)
 );
